FEAT: Sistema completo de decimales para m² + Agregar Folio
✅ **CAMBIOS EN BASE DE DATOS:**
- Migración: m2 BIGINT → DECIMAL(15,2) en planos_folios y planos
- Ejecutada exitosamente sin pérdida de datos

✅ **CORRECCIONES IMPORTADORES:**
- PlanoHistoricoController: intval() → floatval() en líneas 192-194 y 478-479
- Ahora importa correctamente decimales desde Excel

✅ **FORMATO DISPLAY:**
- Hectáreas: 80,33 ha (coma decimal)
- M²: 6.629,22 m² (punto miles + coma decimal)
- Actualizado en todas las vistas y modales

✅ **AGREGAR FOLIO - COMPLETO:**
- Modal con tabs (Quitar/Agregar)
- Radio buttons: Rural (Hijuela) / Urbano (Sitio)
- Conversión automática ha ↔ m² en tiempo real
- Formato chileno: 7,22 ha = 72.200,00 m²
- Búsqueda opcional en Matrix con autocomplete
- Validaciones completas
- Backend: PlanoController::agregarFolio()
- Ruta: POST /planos/{id}/agregar-folio

✅ **API MATRIX BÚSQUEDA:**
- Ruta: GET /api/matrix/buscar
- Método: PlanoImportacionController::buscarFolioMatrix()
- Auto-completa: folio, nombres, apellidos

✅ **VALIDACIONES ACTUALIZADAS:**
- Acepta decimales en todos los inputs m²
- Hectáreas solo para HIJUELA (Rural)
- M² obligatorio siempre

🎯 **RESULTADO:**
- Sistema completo de gestión folios (Agregar/Quitar)
- Decimales funcionando en toda la aplicación
- Importadores corregidos para futuras cargas
- Display formato chileno consistente

+++ app/Http/Controllers/Admin/PlanoHistoricoController.php
@@ -189,9 +189,9 @@ class PlanoHistoricoController extends Controller
             'COMUNA' => $fila[8] ?? '',
             'HIJ' => intval($fila[9] ?? 0),
             'HA' => floatval($fila[10] ?? 0),
-            'M²_HIJ' => intval($fila[11] ?? 0), // Primera columna M²
+            'M²_HIJ' => floatval($fila[11] ?? 0), // Primera columna M² - Con decimales
             'SITIO' => intval($fila[12] ?? 0),
-            'M²_SITIO' => intval($fila[13] ?? 0), // Segunda columna M²
+            'M²_SITIO' => floatval($fila[13] ?? 0), // Segunda columna M² - Con decimales
             'FECHA' => $fila[14] ?? '',
             'AÑO' => intval($fila[15] ?? 0),
             'Responsable' => $fila[16] ?? '',
@@ -475,8 +475,8 @@ class PlanoHistoricoController extends Controller
     private function determinarTipoInmueble($fila)
     {
         $hectareas = floatval($fila['HA'] ?? 0);
-        $m2Hij = intval($fila['M²_HIJ'] ?? 0);
-        $m2Sitio = intval($fila['M²_SITIO'] ?? 0);
+        $m2Hij = floatval($fila['M²_HIJ'] ?? 0);
+        $m2Sitio = floatval($fila['M²_SITIO'] ?? 0);
 
         if ($hectareas > 0 || $m2Hij > 0) {
             // Es HIJUELA
